[React portfolio]
- Code refactored
- Added cv as project in projects section
- Added assets for new project
