+++ src/components/Projects/Projects.scss
@@ -1,0 +1,170 @@
+.projects {
+
+  @media (max-width: 767px) {
+    width: 100%;
+  }
+
+  @media (min-width: 768px) {
+    width: 65%;
+  }
+
+  &__list {
+    width: 100%;
+    padding: 15px;
+    display: flex;
+    flex-wrap: wrap;
+    justify-content: center;
+    align-items: center;
+
+    @media (max-width: 767px) {
+      flex-direction: column;
+    }
+  }
+}
+
+.project {
+  padding: 20px;
+  position: relative;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+
+  &.cv {
+    .project__img {
+      background-image: url('./../../assets/cv.svg');
+    }
+
+    .project__type {
+      &::before {
+        background-image: url('./../../assets/responsive.svg');
+      }
+    }
+  }
+
+  &.star-wars {
+    .project__img {
+      background-image: url('./../../assets/star-wars.svg');
+    }
+
+    .project__type {
+      &::before {
+        background-image: url('./../../assets/desktop.svg');
+      }
+    }
+  }
+
+  @media (max-width: 767px) {
+    width: 100%;
+  }
+
+  @media (min-width: 768px) and (max-width: 1019px) {
+    width: 40%;
+  }
+
+  @media (min-width: 1020px) {
+    width: 50%;
+  }
+
+  &__type {
+    margin-top: 5px;
+    position: relative;
+
+    &::before {
+      content: "";
+      position: absolute;
+      width: 20px;
+      height: 20px;
+      background-position: center;
+      background-size: contain;
+      left: -25px;
+      top: 50%;
+      transform: translateY(-50%);
+    }
+
+    &.desktop {
+      &::before {
+        background-image: url('./../../assets/desktop.svg');
+      }
+    }
+
+    &.responsive {
+      &::before {
+        background-image: url('./../../assets/responsive.svg');
+      }
+    }
+  }
+
+  &__img {
+    background-size: contain;
+    background-position: center;
+    background-repeat: no-repeat;
+
+    @media (max-width: 767px) {
+      width: 100px;
+      height: 100px;
+    }
+
+    @media (min-width: 768px) and (max-width: 1019px) {
+      width: 130px;
+      height: 130px;
+    }
+
+    @media (min-width: 1020px) {
+      width: 160px;
+      height: 160px;
+    }
+
+    &.star-wars {
+      background-image: url('./../../assets/star-wars.svg');
+    }
+
+    &.cv {
+      background-image: url('./../../assets/cv.svg');
+    }
+  }
+
+  &__name {
+    display: block;
+    margin-top: 5px;
+    text-align: center;
+
+    @media (max-width: 767px) {
+      font-size: 18px;
+    }
+
+
+    @media (min-width: 768px) {
+      font-size: 22px;
+    }
+  }
+}
+
+.project-menu {
+  width: 100%;
+  margin-top: 15px;
+  display: flex;
+  justify-content: center;
+
+  &__item {
+    flex: 1 auto;
+    max-width: 50%;
+    text-align: center;
+    border-radius: 5px;
+    border: 2px solid #8d8b8d;
+
+    &+& {
+      margin-left: 10px;
+    }
+  }
+
+  &__a {
+    display: inline-block;
+    color: #000;
+    font-weight: 700;
+    text-transform: uppercase;
+    text-decoration: none;
+    padding: 5px;
+    cursor: pointer;
+  }
+}
