+++ src/components/ScrollToTop/ScrollToTop.scss
@@ -1,5 +1,6 @@
 .scroll-top {
   all: unset;
+  z-index: 2;
   background-color: #fff;
   border: 1px solid #eeedee;
   border-radius: 50%;
@@ -7,7 +8,6 @@
   background-position: center;
   background-size: contain;
   position: fixed;
-  z-index: 999;
   box-shadow: 0 0 5px 1px #c0c0c0;
   transition: all .4s;
   cursor: pointer;
