+++ src/components/Skills/Skills.scss
@@ -1,0 +1,70 @@
+.skills {
+  width: 100%;
+  display: flex;
+  justify-content: center;
+  flex-wrap: wrap;
+  padding: 15px;
+
+  &__list {
+    padding: 30px 0;
+    display: grid;
+    grid-template-columns: repeat(2, 1fr);
+    grid-template-rows: repeat(2, 1fr);
+    grid-column-gap: 5px;
+    grid-row-gap: 20px;
+    background-color: #fff;
+    border: 1px solid #eeedee;
+    border-radius: 15px;
+
+    @media (max-width: 767px) {
+      width: 100%;
+    }
+
+    @media (min-width: 768px) {
+      width: 65%;
+    }
+  }
+}
+
+.skill {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  position: relative;
+
+  @media (min-width: 1020px) {
+    font-size: 20px;
+  }
+
+  &__level {
+    color: #8d8b8d;
+  }
+
+  &__row {
+    display: flex;
+    align-items: center;
+  }
+
+  &__title {
+    position: relative;
+
+    &::after {
+      content: "";
+      position: absolute;
+      background-image: url('./../../assets/verified.svg');
+      background-size: contain;
+
+      @media (max-width: 1019px) {
+        width: 15px;
+        height: 15px;
+        right: -20px;
+      }
+
+      @media (min-width: 1020px) {
+        width: 20px;
+        height: 20px;
+        right: -25px;
+      }
+    }
+  }
+}
